feat(evm): EIP-4788 + blob gas de bloque — blockchain_test 21 158/21 158 en Paris..Cancun
El eje `blockchain_test` suma Cancun completo: 3 473 -> 21 158 / 21 158,
`FAIL 0`. `state_test` intacto en 39 025/39 025 y el diferencial vs revm en
315/315.

Tres piezas:

1. `crates/evm/src/system_call.rs` (modulo nuevo) — la system call de EIP-4788.
   Caller SYSTEM_ADDRESS, 30 M de gas, sin prewarm (el mensaje de sistema
   arranca con el accessed set vacio), sin transferir value, sin bumpear nonce,
   sin receipt y sin sumar al `gasUsed` del bloque. El codigo sale de `code_of`
   crudo, sin resolver delegaciones de EIP-7702: un contrato de sistema
   delegado correria el codigo de otro.

   La orquesta el driver y no `begin_block`: el seam `Vm` esta vendoreado y
   `begin_block(env, state)` no lleva el beacon root — el trait expone
   `system_call_in_block` justamente para que el cliente, que es el que lee el
   header, dispare la llamada.

2. `blockchain/header.rs::check_blob_gas` — las tres reglas de blob gas del
   header: `excessBlobGas` acumulado del padre, `blobGasUsed` contra las txs
   del bloque, y el tope de bloque. El tope es de BLOQUE y no de tx: de los 28
   fixtures de `test_invalid_block_blob_count`, una sola es una tx de 7 blobs;
   las otras 27 reparten 7 blobs entre varias txs de <= 6.

3. `encode_eip4844` — el envelope RLP de tx tipo 3, que ahora ejercitan 1 264
   casos. El tipo 4 sigue fail-closed: no existe antes de Prague.

Los parametros de blob se unifican en una tabla por fork (`blob::blob_params`)
que consumen las tres puertas — validacion de tx, header y precio — con la fila
de Prague (EIP-7691) desde el dia uno: una constante hardcodeada a Cancun pasa
todo este scope en verde y rompe el siguiente en silencio.

Correccion de tipo destapada por el corpus: `blob_base_fee` desbordaba `u64`.
Con un `excessBlobGas` de 1 130 blobs el precio pasa de `u64::MAX` y 13 casos
VALIDOS morian con un error interno. `BLOBBASEFEE` apila 256 bits, asi que el
tipo angosto nunca vino de la EVM; pasa a `u128` en `blob.rs` y en
`interpreter::host::BlockEnv`.

Un bloque publicado solo como `rlp` crudo —sin `blockHeader` ni
`rlp_decoded.blockHeader`— se rechaza como invalido si y solo si el fixture lo
declara asi: EEST omite el cuerpo decodificado justamente cuando el bloque no
decodifica (una tx tipo 3 con `to = None` no es RLP-representable). Sin
`expectException` sigue siendo falla de parseo.

5 tests de integracion nuevos (`crates/evm/tests/system_call.rs`) para lo que
el corpus de Cancun no puede medir: el no-op sin codigo, el fallo de la system
call, y que SYSTEM_ADDRESS no quede en el trie.

+++ crates/interpreter/src/host.rs
@@ -40,7 +40,11 @@ pub struct BlockEnv {
     pub prevrandao: B256,
     /// EIP-7516 (BLOBBASEFEE) / EIP-4844 `fake_exponential`: ya calculado por
     /// quien arma el frame (`evm`); el intérprete solo lo apila.
-    pub blob_base_fee: u64,
+    ///
+    /// `u128` y no `u64`: con un `excessBlobGas` alto —pero alcanzable, y
+    /// presente en el set de EEST— el precio se pasa de `u64::MAX`. El opcode
+    /// apila una palabra de 256 bits; el tipo angosto nunca fue de la EVM.
+    pub blob_base_fee: u128,
 }
 
 /// Datos de la tx que ORIGIN/GASPRICE/BLOBHASH necesitan. Vive en
